脚本

+++ mallplus-portal/src/main/java/com/zscat/mallplus/oms/service/impl/OmsOrderServiceImpl.java
@@ -201,12 +201,16 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
 
     @Override
     public Object preGroupActivityOrder(OrderParam orderParam) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         SmsGroupActivity smsGroupActivity = smsGroupActivityService.getById(orderParam.getGroupId());
         if (ValidatorUtils.notEmpty(smsGroupActivity.getGoodsIds())) {
             List<PmsProduct> productList = (List<PmsProduct>) productService.listByIds(
                     Arrays.asList(smsGroupActivity.getGoodsIds().split(",")).stream().map(s -> Long.parseLong(s.trim())).collect(Collectors.toList()));
             if (productList != null && productList.size() > 0) {
-                UmsMember currentMember = memberService.getNewCurrentMember();
+
                 ConfirmOrderResult result = new ConfirmOrderResult();
                 // 邮费
                 BigDecimal transFee = BigDecimal.ZERO;
@@ -347,14 +351,18 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
      * @return
      */
     @Override
-    public ConfirmOrderResult submitPreview(OrderParam orderParam) {
+    public Object submitPreview(OrderParam orderParam) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         if (ValidatorUtils.empty(orderParam.getTotal())) {
             orderParam.setTotal(1);
         }
         String type = orderParam.getType();
         StopWatch stopWatch = new StopWatch("下单orderType=" + orderParam.getOrderType());
         stopWatch.start("1. 获取购物车商品");
-        UmsMember currentMember = memberService.getNewCurrentMember();
+
         List<OmsCartItem> list = new ArrayList<>();
         if ("3".equals(type)) { // 1 商品详情 2 勾选购物车 3全部购物车的商品
             list = cartItemService.list(currentMember.getId(), null);
@@ -469,14 +477,17 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
      * @return
      */
     @Override
-    public ConfirmListOrderResult submitStorePreview(OrderParam orderParam) {
+    public Object submitStorePreview(OrderParam orderParam) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         ConfirmListOrderResult confirmListOrderResult = new ConfirmListOrderResult();
         List<ConfirmOrderResult> confirmOrderResultList = new ArrayList<>();
         if (ValidatorUtils.empty(orderParam.getTotal())) {
             orderParam.setTotal(1);
         }
 
-        UmsMember currentMember = memberService.getNewCurrentMember();
         List<OmsCartItem> list = new ArrayList<>();
 
         String cart_id_list1 = orderParam.getCartIds();
@@ -564,6 +575,9 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
 
         String type = orderParam.getType();
         UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         OmsOrder order = new OmsOrder();
         // 1. 获取购物车商品
         List<OmsCartItem> cartPromotionItemList = new ArrayList<>();
@@ -836,6 +850,10 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
 
     @Override
     public Object quitGroup(Long id) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         groupMemberMapper.deleteById(id);
 
         return new CommonResult().success("下单成功");
@@ -905,9 +923,8 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
     public CommonResult acceptGroup(OrderParam orderParam) {
 
         OmsOrder order = new OmsOrder();
-
         UmsMember currentMember = memberService.getNewCurrentMember();
-        if (currentMember == null) {
+        if (currentMember==null || currentMember.getId()==null){
             return new CommonResult().fail(100);
         }
         List<OmsCartItem> list = new ArrayList<>();
@@ -1182,6 +1199,10 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Object confimDelivery(Long id) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         OmsOrder order = this.orderMapper.selectById(id);
         if (order.getStatus() != OrderStatus.DELIVERED.getValue()) {
             return new CommonResult().paramFailed("已发货订单才能确认收货");
@@ -1196,6 +1217,10 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
     @Override
     @Transactional(rollbackFor = Exception.class)
     public Object applyRefund(Long id) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         OmsOrder order = this.orderMapper.selectById(id);
         try {
             if (order.getStatus() > 9) {
@@ -1302,10 +1327,14 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
 
     @Transactional
     @Override
-    public OmsOrder blancePay(PayParam payParam) {
+    public Object blancePay(PayParam payParam) {
+
         OmsOrder order = orderMapper.selectById(payParam.getOrderId());
-        UmsMember userDO1 = memberService.getNewCurrentMember();
-        UmsMember userDO = memberService.getById(userDO1.getId());
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
+        UmsMember userDO = memberService.getById(currentMember.getId());
         if (order.getPayAmount().compareTo(userDO.getBlance()) > 0) {
             throw new ApiMallPlusException("余额不足！");
         }
@@ -1370,9 +1399,12 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
 
     @Transactional
     @Override
-    public OmsOrder blancePay(OmsOrder order) {
-        UmsMember userDO1 = memberService.getNewCurrentMember();
-        UmsMember userDO = memberService.getById(userDO1.getId());
+    public Object blancePay(OmsOrder order) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
+        UmsMember userDO = memberService.getById(currentMember.getId());
         if (order.getPayAmount().compareTo(userDO.getBlance()) > 0) {
             throw new ApiMallPlusException("余额不足！");
         }
@@ -1450,6 +1482,10 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
     @Override
     @Transactional
     public Object jifenPay(OrderParam orderParam) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         UmsMember member = memberService.getById(memberService.getNewCurrentMember().getId());
         PmsGifts gifts = giftsService.getById(orderParam.getGoodsId());
         if (gifts.getPrice().intValue() > member.getIntegration()) {
@@ -2138,6 +2174,10 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
 
     @Override
     public Object addGroup(OrderParam orderParam) {
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         List<OmsCartItem> list = new ArrayList<>();
         if (ValidatorUtils.empty(orderParam.getTotal())) {
             orderParam.setTotal(1);
@@ -2194,7 +2234,7 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
     @Override
     public Object addCart(CartParam cartParam) {
         UmsMember currentMember = memberService.getNewCurrentMember();
-        if (ValidatorUtils.empty(currentMember)) {
+        if (currentMember==null || currentMember.getId()==null){
             return new CommonResult().fail(100);
         }
         if (ValidatorUtils.empty(cartParam.getTotal())) {
@@ -2205,8 +2245,7 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
         Integer memberRate = 0; // 会员折扣
         UmsMemberLevel memberLevel = new UmsMemberLevel();
         if (pmsProduct.getIsVip() != null && pmsProduct.getIsVip() == 1) {
-            UmsMember member = memberService.getNewCurrentMember();
-            memberLevel = memberLevelService.getById(member.getMemberLevelId());
+            memberLevel = memberLevelService.getById(currentMember.getMemberLevelId());
             if (memberLevel != null && memberLevel.getPriviledgeMemberPrice() > 0) {
                 memberRate = memberLevel.getPriviledgeMemberPrice();
             }
@@ -2321,6 +2360,9 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
         }
         String type = orderParam.getType();
         UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
 
         // 1. 获取购物车商品
         List<OmsCartItem> cartPromotionItemList = new ArrayList<>();
@@ -2432,7 +2474,10 @@ public class OmsOrderServiceImpl extends ServiceImpl<OmsOrderMapper, OmsOrder> i
     @Transactional
     @Override
     public Object applyRe(ApplyRefundVo itemss) {
-
+        UmsMember currentMember = memberService.getNewCurrentMember();
+        if (currentMember==null || currentMember.getId()==null){
+            return new CommonResult().fail(100);
+        }
         try {
             if (ValidatorUtils.empty(itemss.getItems()) ||itemss.getItems().split(",").length<1){
                 return new CommonResult().failed("添加失败");
